scrapi/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
boto3==1.14.47
botocore==1.17.47
docutils==0.15.2
jmespath==0.10.0
s3transfer==0.3.3